+++ src/bin.rs
@@ -31,7 +31,7 @@ mod pathfinding;
 mod pigeon;
 mod position;
 mod settings;
-mod sort;
+mod depthsort;
 mod sprite;
 mod tilemap;
 mod view;
@@ -41,7 +41,7 @@ use common::DeltaTime;
 use grid::{Grid, GridPosition};
 use pathfinding::{components::Pather, systems::PathfindingSystem, AStar, Locomotion, GROUND_WALK};
 use position::Position;
-use sort::{DepthBuffer, IsometricSorter};
+use depthsort::{DepthBuffer, IsometricSorter};
 use sprite::{OnRender, Sprite, SpriteRenderer};
 use tilemap::{Tile, TileObj, Tilemap};
 use view::components::IsometricCamera;
